Use TX Power slider to correct data plot
Previously, the path loss for the measured data was incorrectly
calculated as the delta between the peak measured power and a given data
point. Instead, correctly calculate path loss by providing an adjustable
TX power (since we are unable to get actual transmitted power info for a
given cell) for better analysis.

Signed-off-by: Trevor Gamblin <tvgamblin@gmail.com>